REFACTOR package structure of onboarding feature
Onboarding-Package-Structure-Fix
