--- feature/onboarding/src/main/kotlin/nl/q42/template/onboarding/presentation/start/OnboardingStartViewState.kt
@@ -1,3 +1,0 @@
-package nl.q42.template.onboarding.presentation.start
-
-class OnboardingStartViewState(val title: String)
