+++ feature/onboarding/src/main/kotlin/nl/q42/template/onboarding/start/presentation/OnboardingStartViewState.kt
@@ -1,0 +1,3 @@
+package nl.q42.template.onboarding.start.presentation
+
+class OnboardingStartViewState(val title: String)
